requirements.txt
Django
djangorestframework
djangorestframework-authtoken